Provided some scripts for pattern matching
The main method can be found in testAhocorasick, the rest is just decorum I used to test and write the script. 
Please note that the script is full of prints, you can remove them if they annoy you. Also there is a bit of code to execute at the end of the file, which is unorthodox to say the least.

If you want to run the tests just run the shell script called demoAndSetup.sh, it will set everything and run the test (with a lot of verbose).
dumpProducer creates two random 100 MB text files to test search substrings in.
unifiedBlacklist.zip contains the blacklist found on http://www.shallalist.de/ and will be extracted by demoAndSetup if required.

Also, even though I think that efficient pattern matching can be a helpful tool for the project in general, I am not certain that this is the best way to use it, as literally any subterfuge will prevent the script from detecting the malicious URL (for instance splitting it in two sub-strings joined with a "+" would be enough). But a least you can see here what the algorithms capabilities are and how it could be useful.

I am not fully satisfied wit how the script has been written, and I will rework it a little tomorrow.